refactor: 가정예배 공지 추가 ui 수정
