+++ apps/admin/src/components/homeworship/index.tsx
@@ -1,0 +1,94 @@
+import { usePostHomeWorship } from "@/query/homeWorship";
+import dynamic from "next/dynamic";
+import { useState } from "react";
+import { SubmitHandler, useForm } from "react-hook-form";
+import { IHomeWorshipForm } from "type";
+import { Spinner } from "ui";
+import { Input } from "../ui/input";
+
+const Editor = dynamic(() => import("@/components/common/editor").then((mod) => mod.Editor), {
+  ssr: false,
+  loading: () => <Spinner />,
+});
+
+const Homeworship = () => {
+  const { register, handleSubmit, reset } = useForm<IHomeWorshipForm>();
+
+  const { mutate, isPending } = usePostHomeWorship();
+
+  const [content, setContent] = useState("");
+
+  const onSubmit: SubmitHandler<IHomeWorshipForm> = async (data) => {
+    const formData = new FormData();
+
+    if (
+      data.image.length === 0 ||
+      !data.date ||
+      !data.title ||
+      !data.password ||
+      !data.userName ||
+      !content
+    )
+      return alert("모든 정보를 입력해주세요.");
+    if (data.image.length !== 1) return alert("사진은 한 장 업로드 가능합니다.");
+
+    formData.append("title", data.title);
+    formData.append("date", data.date);
+    formData.append("content", content);
+    formData.append("password", data.password);
+    formData.append("userName", data.userName);
+    formData.append("isPinned", "checked");
+
+    Array.from(data.image).forEach((image) => {
+      formData.append(`image-file`, image);
+      formData.append(`image-name`, image.name);
+    });
+
+    mutate(formData, {
+      onSuccess: () => {
+        alert("추가되었습니다.");
+        reset();
+      },
+      onError: () => {
+        alert("에러가 발생했습니다. 다시 시도해주세요.");
+      },
+    });
+  };
+
+  return (
+    <form className="flex flex-col gap-3" onSubmit={handleSubmit(onSubmit)}>
+      <label className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">예배 날짜</p>
+        <Input className="w-[233px]" type="date" {...register("date")} />
+      </label>
+      <label className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">제목</p>
+        <Input className="w-[233px]" {...register("title")} />
+      </label>
+      <label className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">사진 업로드</p>
+        <Input className="w-[233px]" type="file" accept="image/*" {...register("image")} />
+      </label>
+      <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">글</p>
+        <Editor setValue={setContent} />
+      </div>
+      <label className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">작성자</p>
+        <Input className="w-[233px]" {...register("userName")} />
+      </label>
+      <label className="flex flex-col gap-2 sm:flex-row sm:items-center sm:gap-5">
+        <p className="text-xl font-bold">비밀번호</p>
+        <Input className="w-[233px]" type="password" {...register("password")} />
+      </label>
+      <button
+        disabled={isPending}
+        className="mt-5 rounded-md bg-blue-500 py-2 font-bold text-white"
+      >
+        공지 올리기
+      </button>
+    </form>
+  );
+};
+
+export default Homeworship;
